Reformatted Setting Page & Assets
