+++ assets/sounds/infected_notification.mp3
