+++ lib/utilities/screen_size_config.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 
 class SizeConfig {
   static MediaQueryData _mediaQueryData = 0 as MediaQueryData;
