+++ pubspec.yaml
@@ -50,6 +50,7 @@ dependencies:
   audioplayers: any
   device_info: ^2.0.3
   device_info_plus: ^3.1.1
+  widget_view: ^0.2.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
